+++ demo/proxy.ts
@@ -2,6 +2,6 @@ import proxy from "../server/proxy.ts";
 
 Deno.serve({
   handler: proxy({
-    args: ["run", "-A", "--unstable-kv", "demo/serve.ts"],
+    serve: "demo/dev.ts",
   }),
 });