+++ demo/serve.ts
@@ -1,6 +1,0 @@
-import app from "./dev.ts";
-
-Deno.serve({
-  handler: app.init().fetch,
-  port: 3000,
-});